geospatial_agents/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20
tavily-python>=0.3.0
geopandas>=0.14.0
rasterio>=1.3.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
folium>=0.15.0
contextily>=1.4.0
matplotlib>=3.8.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
huggingface_hub[cli]>=0.20.0
huggingface_hub>=0.20.0
datasets>=2.14.0
sentinelhub>=3.10.0
